requirements.txt
texttable
base58
ecdsa
python-dateutil
whichcraft
termcolor
PyYAML
docker
docker-compose
netifaces
psutil
six
ruamel.ordereddict
ruamel.yaml
future
duckietown-challenges-daffy>=5.1.4
zeroconf
dt-authentication-daffy
dt-data-api-daffy>=1.1.1
pytz
dockertown>=0.2.2
dtproject>=0.0.5

questionary
pyfiglet
termcolor
